+++ packages/form/src/demos/modal-form.tsx
@@ -29,6 +29,7 @@ export default () => {
         await waitTime(2000);
         console.log(values);
         message.success('提交成功！');
+        return true;
       }}
     >
       <ProForm.Group>