+++ packages/layout/src/PageContainer/index.tsx
@@ -1,8 +1,9 @@
-import { PageHeader, Tabs, Affix } from 'antd';
+import { PageHeader, Tabs, Affix, ConfigProvider } from 'antd';
 import React, { useContext, ReactNode } from 'react';
 import classNames from 'classnames';
 import { TabsProps, TabPaneProps } from 'antd/lib/tabs';
 import { PageHeaderProps } from 'antd/lib/page-header';
+import { AffixProps } from 'antd/lib/affix';
 
 import RouteContext, { RouteContextType } from '../RouteContext';
 import GridContent from '../GridContent';
@@ -24,9 +25,24 @@ export interface PageContainerProps extends PageHeaderTabConfig, Omit<PageHeader
   extraContent?: React.ReactNode;
   prefixCls?: string;
   footer?: ReactNode[];
+
+  /**
+   * @name 是否显示背景色
+   */
   ghost?: boolean;
+
+  /**
+   * @name PageHeader 的配置
+   * @description 与 antd 完全相同
+   */
   header?: PageHeaderProps;
   pageHeaderRender?: (props: PageContainerProps) => React.ReactNode;
+
+  /**
+   * @name 固钉的配置
+   * @description 与 antd 完全相同
+   */
+  affixProps?: AffixProps;
 }
 
 /**
@@ -125,8 +141,12 @@ const defaultPageHeaderRender = (
 };
 
 const PageContainer: React.FC<PageContainerProps> = (props) => {
-  const { children, style, footer, ghost, fixedHeader, prefixCls = 'ant-pro' } = props;
+  const { children, style, footer, affixProps, ghost, fixedHeader } = props;
   const value = useContext(RouteContext);
+
+  const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
+  const prefixCls = props.prefixCls || getPrefixCls('pro');
+
   const prefixedClassName = `${prefixCls}-page-container`;
 
   const className = classNames(prefixedClassName, props.className, {
@@ -146,7 +166,13 @@ const PageContainer: React.FC<PageContainerProps> = (props) => {
   return (
     <div style={style} className={className}>
       {fixedHeader ? (
-        <Affix offsetTop={value.hasHeader ? value.headerHeight : 0}>{headerDom}</Affix>
+        // 在 hasHeader 且 fixedHeader 的情况下，才需要设置高度
+        <Affix
+          offsetTop={value.hasHeader && value.fixedHeader ? value.headerHeight : 0}
+          {...affixProps}
+        >
+          {headerDom}
+        </Affix>
       ) : (
         headerDom
       )}
@@ -165,7 +191,7 @@ const PageContainer: React.FC<PageContainerProps> = (props) => {
           </div>
         ) : null}
       </GridContent>
-      {footer && <FooterToolbar>{footer}</FooterToolbar>}
+      {footer && <FooterToolbar prefixCls={prefixCls}>{footer}</FooterToolbar>}
     </div>
   );
 };