+++ packages/table/src/demos/renderTable.tsx
@@ -1,6 +1,17 @@
-import React, { useState, useEffect } from 'react';
-import { Result, Card, Descriptions } from 'antd';
+import React, { useState } from 'react';
+import { Card, Menu, Descriptions } from 'antd';
 import ProTable, { ProColumns } from '@ant-design/pro-table';
+import { AppstoreOutlined, MailOutlined, SettingOutlined } from '@ant-design/icons';
+
+const waitTime = (time: number = 100) => {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve(true);
+    }, time);
+  });
+};
+
+const { SubMenu } = Menu;
 
 export interface TableListItem {
   key: number;
@@ -29,7 +40,6 @@ const columns: ProColumns<TableListItem>[] = [
   {
     title: '更新时间',
     key: 'since2',
-    width: 120,
     dataIndex: 'createdAt',
     valueType: 'date',
   },
@@ -41,14 +51,7 @@ const columns: ProColumns<TableListItem>[] = [
 ];
 
 export default () => {
-  const [loading, setLoading] = useState(true);
-  const [dataSource, setDataSource] = useState<TableListItem[]>([]);
-  useEffect(() => {
-    setTimeout(() => {
-      setLoading(false);
-      setDataSource(tableListDataSource);
-    }, 5000);
-  }, []);
+  const [key, setKey] = useState('1');
 
   return (
     <ProTable<TableListItem>
@@ -64,7 +67,54 @@ export default () => {
             width: '100%',
           }}
         >
-          <Result status="404" title="404" subTitle="404" />
+          <Menu
+            onSelect={(e) => setKey(e.key as string)}
+            style={{ width: 256 }}
+            defaultSelectedKeys={['1']}
+            defaultOpenKeys={['sub1']}
+            mode="inline"
+          >
+            <Menu.SubMenu
+              key="sub1"
+              title={
+                <span>
+                  <MailOutlined />
+                  <span>Navigation One</span>
+                </span>
+              }
+            >
+              <Menu.ItemGroup key="g1" title="Item 1">
+                <Menu.Item key="1">Option 1</Menu.Item>
+                <Menu.Item key="2">Option 2</Menu.Item>
+              </Menu.ItemGroup>
+              <Menu.ItemGroup key="g2" title="Item 2">
+                <Menu.Item key="3">Option 3</Menu.Item>
+                <Menu.Item key="4">Option 4</Menu.Item>
+              </Menu.ItemGroup>
+            </Menu.SubMenu>
+            <SubMenu key="sub2" icon={<AppstoreOutlined />} title="Navigation Two">
+              <Menu.Item key="5">Option 5</Menu.Item>
+              <Menu.Item key="6">Option 6</Menu.Item>
+              <SubMenu key="sub3" title="Submenu">
+                <Menu.Item key="7">Option 7</Menu.Item>
+                <Menu.Item key="8">Option 8</Menu.Item>
+              </SubMenu>
+            </SubMenu>
+            <SubMenu
+              key="sub4"
+              title={
+                <span>
+                  <SettingOutlined />
+                  <span>Navigation Three</span>
+                </span>
+              }
+            >
+              <Menu.Item key="9">Option 9</Menu.Item>
+              <Menu.Item key="10">Option 10</Menu.Item>
+              <Menu.Item key="11">Option 11</Menu.Item>
+              <Menu.Item key="12">Option 12</Menu.Item>
+            </SubMenu>
+          </Menu>
           <div
             style={{
               flex: 1,
@@ -90,15 +140,15 @@ export default () => {
           </Descriptions>
         </Card>
       )}
-      loading={loading}
-      dataSource={dataSource}
-      options={{
-        reload: () => {
-          setLoading(true);
-          setTimeout(() => {
-            setLoading(false);
-          }, 1000);
-        },
+      params={{
+        key,
+      }}
+      request={async () => {
+        await waitTime(200);
+        return {
+          success: true,
+          data: tableListDataSource,
+        };
       }}
       dateFormatter="string"
       headerTitle="自定义 table"