+++ scripts/utils/exec.js
@@ -2,7 +2,12 @@ const { spawn } = require('child_process');
 
 module.exports = function exec(command, args, opts) {
   return new Promise((resolve, reject) => {
-    const child = spawn(command, args, Object.assign({ stdio: 'inherit', env: process.env }, opts));
+    const child = spawn(command, args, {
+      shell: true,
+      stdio: 'inherit',
+      env: process.env,
+      ...opts,
+    });
     child.once('error', (err) => {
       console.log(err);
       reject(err);