+++ packages/descriptions/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.25](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.0.24...@ant-design/pro-descriptions@1.0.25) (2020-10-19)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.0.24](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.0.23...@ant-design/pro-descriptions@1.0.24) (2020-10-15)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.0.23](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.0.22...@ant-design/pro-descriptions@1.0.23) (2020-10-15)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
 ## [1.0.22](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.0.20...@ant-design/pro-descriptions@1.0.22) (2020-10-12)
 
 **Note:** Version bump only for package @ant-design/pro-descriptions